1202 opprydding og fiksing av todos
