+++ src/main/kotlin/no/nav/tiltakspenger/datadeling/domene/Kilde.kt
@@ -1,0 +1,8 @@
+package no.nav.tiltakspenger.datadeling.domene
+
+enum class Kilde(
+    val navn: String,
+) {
+    TPSAK("tp"),
+    ARENA("arena"),
+}
