+++ src/main/kotlin/no/nav/tiltakspenger/datadeling/domene/PeriodisertKilde.kt
@@ -4,5 +4,5 @@ import no.nav.tiltakspenger.libs.periodisering.Periode
 
 data class PeriodisertKilde(
     val periode: Periode,
-    val kilde: String,
+    val kilde: Kilde,
 )
