+++ src/main/kotlin/no/nav/tiltakspenger/datadeling/domene/Vedtak.kt
@@ -13,7 +13,7 @@ data class Vedtak(
     val vedtakId: String,
     val sakId: String,
     val saksnummer: String?,
-    val kilde: String,
+    val kilde: Kilde,
     val fnr: Fnr,
 )
