+++ src/main/kotlin/no/nav/tiltakspenger/datadeling/service/VedtakService.kt
@@ -4,11 +4,12 @@ import arrow.core.Either
 import arrow.core.left
 import arrow.core.right
 import no.nav.tiltakspenger.datadeling.client.arena.ArenaClient
+import no.nav.tiltakspenger.datadeling.domene.Kilde
 import no.nav.tiltakspenger.datadeling.domene.PeriodisertKilde
 import no.nav.tiltakspenger.datadeling.domene.Systembruker
 import no.nav.tiltakspenger.datadeling.domene.Systembrukerrolle
 import no.nav.tiltakspenger.datadeling.domene.TiltakspengerVedtak
-import no.nav.tiltakspenger.datadeling.motta.app.VedtakRepo
+import no.nav.tiltakspenger.datadeling.motta.infra.db.VedtakRepo
 import no.nav.tiltakspenger.libs.common.Fnr
 import no.nav.tiltakspenger.libs.periodisering.Periode
 import no.nav.tiltakspenger.libs.periodisering.Periodisering
@@ -18,7 +19,7 @@ class VedtakService(
     private val vedtakRepo: VedtakRepo,
     private val arenaClient: ArenaClient,
 ) {
-    suspend fun hentTpVedtak(
+    fun hentTpVedtak(
         fnr: Fnr,
         periode: Periode,
         systembruker: Systembruker,
@@ -29,7 +30,7 @@ class VedtakService(
                 harRollene = systembruker.roller.toList(),
             ).left()
         }
-        return vedtakRepo.hentForFnrOgPeriode(fnr, periode, "tp").toTidslinje().right()
+        return vedtakRepo.hentForFnrOgPeriode(fnr, periode, Kilde.TPSAK).toTidslinje().right()
     }
 
     suspend fun hentPerioder(
@@ -43,9 +44,9 @@ class VedtakService(
                 harRollene = systembruker.roller.toList(),
             ).left()
         }
-        val vedtak = vedtakRepo.hentForFnrOgPeriode(fnr, periode, "tp").map { vedtak ->
+        val vedtak = vedtakRepo.hentForFnrOgPeriode(fnr, periode, Kilde.TPSAK).map { vedtak ->
             PeriodisertKilde(
-                kilde = "tp",
+                kilde = vedtak.kilde,
                 periode = vedtak.periode,
             )
         }
