+++ src/test/kotlin/no/nav/tiltakspenger/datadeling/felles/TestDataHelper.kt
@@ -1,8 +1,8 @@
 package no.nav.tiltakspenger.datadeling.felles
 
 import mu.KotlinLogging
-import no.nav.tiltakspenger.datadeling.motta.infra.db.BehandlingPostgresRepo
-import no.nav.tiltakspenger.datadeling.motta.infra.db.VedtakPostgresRepo
+import no.nav.tiltakspenger.datadeling.motta.infra.db.BehandlingRepo
+import no.nav.tiltakspenger.datadeling.motta.infra.db.VedtakRepo
 import no.nav.tiltakspenger.libs.persistering.infrastruktur.PostgresSessionFactory
 import no.nav.tiltakspenger.libs.persistering.infrastruktur.SessionCounter
 import javax.sql.DataSource
@@ -13,8 +13,8 @@ internal class TestDataHelper(
     private val log = KotlinLogging.logger {}
     private val sessionCounter = SessionCounter(log)
     val sessionFactory = PostgresSessionFactory(dataSource, sessionCounter)
-    val vedtakPostgresRepo = VedtakPostgresRepo(sessionFactory)
-    val behandlingPostgresRepo = BehandlingPostgresRepo(sessionFactory)
+    val vedtakRepo = VedtakRepo(sessionFactory)
+    val behandlingRepo = BehandlingRepo(sessionFactory)
 }
 
 private val dbManager = TestDatabaseManager()
